[Edit]admin

+++ app/src/main/java/sa/ksu/gpa/saleem/Admin/AdminAdvices.kt
@@ -1,14 +1,26 @@
 package sa.ksu.gpa.saleem.Admin
 
+import android.content.Intent
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import android.view.LayoutInflater
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import cn.pedant.SweetAlert.SweetAlertDialog
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.SetOptions
 import kotlinx.android.synthetic.main.activity_admin_advices.*
+import kotlinx.android.synthetic.main.advice_dialog.view.*
+import sa.ksu.gpa.saleem.MyAdvice
 import sa.ksu.gpa.saleem.R
+import sa.ksu.gpa.saleem.loginn
 import java.text.SimpleDateFormat
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
@@ -19,16 +31,18 @@ class AdminAdvices : AppCompatActivity() {
     lateinit var adapter: ReportedAdvicesAdapter
     private lateinit var db: FirebaseFirestore
     lateinit var recyclerView: RecyclerView
-    var cal : Double = 0.0
-    var key_list:ArrayList<String> = ArrayList()
-    var list:ArrayList<ReportedAdvices> = ArrayList()
+    lateinit var adviceID: String
+    lateinit var title: String
+    var cal: Double = 0.0
+    var key_list: ArrayList<String> = ArrayList()
+    var list: ArrayList<ReportedAdvices> = ArrayList()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_admin_advices)
         db = FirebaseFirestore.getInstance()
         recyclerView = findViewById(R.id.recyclerViewRepAdv)
-        rep_adv_back_button.setOnClickListener{
+        rep_adv_back_button.setOnClickListener {
             onBackPressed()
         }
         getReportedAdvices()
@@ -36,28 +50,28 @@ class AdminAdvices : AppCompatActivity() {
 
     private fun getReportedAdvices() {
         db.collection("ReportedAdvices").get()
-            .addOnSuccessListener{ documents ->
-                for(document in documents){
+            .addOnSuccessListener { documents ->
+                for (document in documents) {
                     key_list.add(document.id)
                     var userID = document.get("reporterUID").toString()
-                    var adviceID = document.get("adviceID").toString()
+                    adviceID = document.get("adviceID").toString()
                     var report = document.get("text").toString()
                     var reportedAdv = ReportedAdvices(userID, adviceID, report)
                     list.add(reportedAdv)
-                    Log.d("ADMIN","List : "+list)
+                    Log.d("ADMIN", "List : " + list)
 
                 }
-                adapter = ReportedAdvicesAdapter(list,  object  : ReportedAdvicesAdapter.OnActionClick {
-                    override fun onClick(item: ReportedAdvices, position: Int) {
-                        showDescItem(item,position)
-                    }
+                adapter =
+                    ReportedAdvicesAdapter(list, object : ReportedAdvicesAdapter.OnActionClick {
+                        override fun onClick(item: ReportedAdvices, position: Int) {
+                            showDescItem(item, position)
+                        }
 
 
-
-                    override fun onDelete(item: ReportedAdvices, position: Int) {
-                        deleteItem(item,position, key_list[position])
-                    }
-                })
+                        override fun onDelete(item: ReportedAdvices, position: Int) {
+                            deleteItem(item, position, key_list[position])
+                        }
+                    })
                 recyclerView.layoutManager = LinearLayoutManager(this)
                 recyclerView.adapter = adapter
             }
@@ -65,7 +79,7 @@ class AdminAdvices : AppCompatActivity() {
 
     }
 
-    fun getCurrentDate():String {
+    fun getCurrentDate(): String {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             val current = LocalDateTime.now()
             val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.ENGLISH)
@@ -78,21 +92,64 @@ class AdminAdvices : AppCompatActivity() {
     }
 
     private fun showDescItem(item: ReportedAdvices, position: Int) {
+        editAdviceDialog(item, key_list[position], position)
         adapter.notifyDataSetChanged()
 
     }
+
     private fun deleteItem(item: ReportedAdvices, position: Int, key: String) {
-        list.removeAt(position)
-        key_list.removeAt(position)
-        adapter.notifyDataSetChanged()
-        db.collection("Advices").document(key).delete()
-            .addOnSuccessListener { Log.d("TAG", "DocumentSnapshot successfully deleted!") }
-            .addOnFailureListener { e -> Log.w("TAG", "Error deleting document", e) }
 
 
+        SweetAlertDialog(this, SweetAlertDialog.WARNING_TYPE)
+            .setTitleText("هل متاكد من حذف النصيحة؟")
+            .setConfirmButton("نعم") { sDialog -> sDialog.dismissWithAnimation()
+                list.removeAt(position)
+                key_list.removeAt(position)
+                adapter.notifyDataSetChanged()
+                db.collection("Advices").document(key).delete()
+                    .addOnSuccessListener { Log.d("TAG", "DocumentSnapshot successfully deleted!") }
+                    .addOnFailureListener { e ->
+                        Log.w("TAG", "Error deleting document", e)
+                    }
+            }.setCancelButton("إلغاء"){
+                it.dismissWithAnimation()
+
+            }
+            .show()
     }
+
 
     private fun showEditItem(item: ReportedAdvices, position: Int) {
         adapter.notifyDataSetChanged()
+    }
+
+    private fun editAdviceDialog(item: ReportedAdvices, key: String, position: Int) {
+
+        db.collection("Advices").whereEqualTo("UID", adviceID).get()
+            .addOnSuccessListener { documents ->
+                for (document in documents) {
+                    key_list.add(document.id)
+                    title = document.get("text").toString()
+                    /*   var adviceID = document.get("adviceID").toString()
+                    var report = document.get("text").toString()*/
+
+                    Log.d("ADMIN", "List : " + list)
+
+                }
+
+                showDialogWithOkButton(title)
+
+            }
+    }
+
+    private fun showDialogWithOkButton(msg: String) {
+        SweetAlertDialog(this, SweetAlertDialog.NORMAL_TYPE)
+            .setTitleText(msg)
+            .setConfirmButton("حسناً") { sDialog ->
+                sDialog.dismissWithAnimation()
+
+
+            }
+            .show()
     }
 }
